requirements.txt
streamlit
azure-storage-blob
pandas
matplotlib
plotly
numpy